Port build to Windows and fix trig functions
Update build scripts for Windows targets and correct math function usage.

- apoplexy.make & raylib.make: change targets to produce Apoplexy.exe / raylib.lib, update OBJDIR naming, switch platform define to -D_WIN32, replace UNIX linker libs with Windows libs (winmm, gdi32, opengl32), add / -mwindows and subsystem linker flags for release builds, and adjust library search paths. These changes prepare the project for Windows/Mingw-style linking instead of Unix.
- Improve Windows clean logic: use a foreach deletion sequence for GENERATED files to avoid failing del /s /q on non-file patterns.
- src/player/player.cpp & src/util/math.cpp: replace std::sinf/std::cosf with std::sin/std::cos (and std::sinf -> std::sin) for consistency and to use the proper overloads.

These edits enable building on Windows toolchains and fix inconsistent use of float trig functions in the codebase.

+++ src/player/player.cpp
@@ -78,7 +78,7 @@ void Player::updateMouseLook(float dt) {
 void Player::updateMovement(float dt, const Level &level) {
   Vector3 moveDir{0.0f, 0.0f, 0.0f};
 
-  Vector3 forward{std::sinf(yaw), 0.0f, std::cosf(yaw)};
+  Vector3 forward{std::sin(yaw), 0.0f, std::cos(yaw)};
 
   Vector3 right{forward.z, 0.0f, -forward.x};
